Used opencsv api and adds more data points
