+++ src/main/webapp/WEB-INF/clinics-data-with-header.csv
@@ -9,3 +9,27 @@ lat,lng,title,address,phoneNum,services
 46.83901247,-96.7670618,Moorhead Clinic,803 Belsly Boulevard,(218) 236-7145,"Abortion Referrals ; Birth Control ; HIV Testing ; Men's Health Care ; Morning-After Pill (Emergency Contraception) ; Pregnancy Testing & Services ; STD Testing, Treatment & Vaccines ; Women's Health Care"
 47.49314144,-94.90623336,Bemidji Clinic,2504 Hannah Ave NW,(218) 751-8683,"Abortion Referrals ; Birth Control ; HIV Testing ; Men's Health Care ; Morning-After Pill (Emergency Contraception) ; Pregnancy Testing & Services ; STD Testing, Treatment & Vaccines ; Women's Health Care"
 43.514517,-96.80785,Sioux Falls Clinic,6511 West 41st Street,(605) 361-5100,"Abortion Services ; Birth Control ; HIV Testing ; Men's Health Care ; Morning-After Pill (Emergency Contraception) ; Pregnancy Testing & Services ; STD Testing, Treatment & Vaccines ; Women's Health Care"
+41.77969238,-87.64142791,Englewood MHC,641 W. 63rd St,312-747-7496,Mental Health Clinic
+41.81632597,-87.60684674,Greater Grand/MID-South MHC ,4314 S. Cottage Grove,312-747-0036,Mental Health Clinic
+41.79327456,-87.72766409,Greater Lawn MHC,4150 W. 55th Street,312-747-1020,Mental Health Clinic
+41.86645597,-87.68901361,Lawndale MHC  ,1201 S. Campbell St,312-746-5905,Mental Health Clinic
+41.98593151,-87.72836921,North River MHC,5801 N. Pulaski Road,312-744-1906,Mental Health Clinic
+41.9022123,-87.74884541,Austin STI Specialty Clinic,"4909 W. Division St., Suite 411",312-746-4871,STI Specialty Clinic
+41.77969238,-87.64142791,Englewood STI Specialty Clinic,"641 W. 63rd St., Lower Level",312-747-8900,STI Specialty Clinic
+41.93436684,-87.64598213,Lakeview STI Specialty Clinic,"2861 N. Clark, 2nd Floor",312-744-5507,STI Specialty Clinic
+41.68548576,-87.61798987,Roseland  STI Specialty Clinic,200 E. 115th St.,312-747-2831,STI Specialty Clinic
+41.84858266,-87.68539419,Alivio Medical Center,2355 S. Western Ave.,773-843-4220,"Women, Infant, Children Clinic"
+41.8541501,-87.6532229,Alivio Medical Center - Morgan,966 W. 21st St.,312-829-6838,"Women, Infant, Children Clinic"
+41.99060139,-87.69073886,"Asian Human Services Family Health Center, Inc.",2424 W. Peterson Ave.,773-761-2324,"Women, Infant, Children Clinic"
+41.9022123,-87.74884541,Austin Clinic,4909 W. Division St.,312-746-4797,"Women, Infant, Children Clinic"
+41.72945712,-87.55247782,Chicago Family Health Center,9119 S. Exchange Ave.,"773-768-0373 ext. 1180, 1181, 1182","Women, Infant, Children Clinic"
+41.77969238,-87.64142791,Englewood NHC,641 W. 63rd St.,312-747-4814,"Women, Infant, Children Clinic"
+41.90311804,-87.68777986,Erie Health Center ,2418 W. Division St.,312-744-8118,"Women, Infant, Children Clinic"
+41.89500555,-87.66981323,Erie Health Center ,1701 W. Superior St.,312-432-7372 ,"Women, Infant, Children Clinic"
+41.79507521,-87.60609661,Friend Family Health Center Inc.,800 E. 55th St.,773-702-3719,"Women, Infant, Children Clinic"
+41.79327456,-87.72766409,Greater Lawn Health Center,4150 W. 55th St.,312-747-5415,"Women, Infant, Children Clinic"
+41.84169281,-87.6250831,Henry Booth House,2907 S. Wabash,312-225-1982 ,"Women, Infant, Children Clinic"
+41.8582821,-87.66601326,Lower West Side NHC,1713 S. Ashland Ave.,312-743-0600,"Women, Infant, Children Clinic"
+41.96536653,-87.65176024,Uptown NHC,845 W. Wilson Ave.,312-744-0073,"Women, Infant, Children Clinic"
+41.84804957,-87.70529466,Westside Health Partnership,2400 S. Kedzie Ave.,773-542-0288,"Women, Infant, Children Clinic"
+41.87739694,-87.62751493,WIC Main Office,"333 S. State St., 2nd Floor",312-747-9140 ,"Women, Infant, Children Clinic"
