+++ src/main/webapp/clinics-map.html
@@ -40,7 +40,7 @@
     <div id="content-box">
     <h1 id="maps-heading">Women's Health Clinics</h1>
     <p id="maps-subheading">Click on the red markers for more information</p>
-    <div id="map"></div>
+    <div id="map">Loading...</div>
     <br>
     </div>
     <br>
